requirements.txt
Flask==1.1.2
machaao==0.3.0
python-dotenv==0.15.0